requirements.txt
Django>=3.0
psycopg2>=2.8.5
gunicorn>=20.0.4
markdown>=3.3
requests>=2.25.1
django-ckeditor>=6.0
beautifulsoup4>=4.9.3
google-api-python-client>=2.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.3
django-allauth>=0.13.0
django-sass-processor>=1.0.1
libsass>=0.21.0
boto3>=1.18.24
django-storages>=1.11.1